Deletion, refactor names
