+++ pyproject.toml
@@ -120,6 +120,6 @@ where = ["src"]
 skip = [".github/*"]
 
 [project.entry-points.'nomad.plugin']
-auto_xrd_schema = "nomad_auto_xrd:schema"
-auto_xrd_app = "nomad_auto_xrd.apps:auto_xrd_models_app"
-auto_xrd_example_upload = "nomad_auto_xrd.example_uploads:auto_xrd_example_upload"
+schema = "nomad_auto_xrd:schema_entry_point"
+models_app = "nomad_auto_xrd.apps:models_app_entry_point"
+example_upload = "nomad_auto_xrd.example_uploads:example_upload_entry_point"
