+++ src/nomad_auto_xrd/example_uploads/__init__.py
@@ -1,6 +1,6 @@
 from nomad.config.models.plugins import ExampleUploadEntryPoint
 
-auto_xrd_example_upload = ExampleUploadEntryPoint(
+example_upload_entry_point = ExampleUploadEntryPoint(
     title='Auto XRD example upload',
     category='Examples',
     description='This example upload contains a notebook and instructions on how to '
