+++ tests/apps/test_app.py
@@ -1,5 +1,5 @@
 def test_importing_app():
     # this will raise an exception if pydantic model validation fails for th app
-    from nomad_auto_xrd.apps import app_entry_point
+    from nomad_auto_xrd.apps import models_app_entry_point
 
-    assert app_entry_point.app.label == 'Auto XRD Models'
+    assert models_app_entry_point.app.label == 'Auto XRD Models'
