+++ tests/data/test.archive.yaml
@@ -1,4 +1,4 @@
 data:
-  m_def: nomad_auto_xrd.schema_packages.auto_xrd.AutoXRDModel
+  m_def: nomad_auto_xrd.schema.AutoXRDModel
   cif_files:
     - 'Cu3PS4_31.cif'
